refactor: remove ValidateUpload and replace it with ValidateDataReceived
