--- src/Middlewares/ValidateUpload.ts
@@ -1,24 +1,0 @@
-import { Request, Response, NextFunction } from 'express';
-import { z } from 'zod';
-import UploadSchema from '@Schemas/UploadSchema';
-import HttpError from '@Helpers/HttpError';
-
-const httpError = new HttpError();
-
-const validateUpload = (req: Request, res: Response, next: NextFunction) => {
-  try {
-    UploadSchema.parse(req.body);
-    next();
-  } catch (error) {
-    if (error instanceof z.ZodError) {
-      const errorMessage = error.errors.map((err) => err.message);
-      return httpError.BadRequest(res, {
-        error_code: 'INVALID_DATA',
-        error_description: errorMessage,
-      });
-    }
-    next(error);
-  }
-};
-
-export default validateUpload;
